m3u8优化完成 支持list

+++ BaseM3U8/src/main/java/com/bihe0832/android/base/m3u8/tools/M3U8Tools.kt
@@ -175,14 +175,18 @@ object M3U8Tools {
     }
 
     fun mergeM3U8(m3u8Dir: String, videoName: String, listener: M3U8Listener) {
-        val finalOutPutFile = File(videoName)
         try {
+            var a = parseIndex("", "", m3u8Dir + "local.m3u8")
+            if(a.tsList.isEmpty()){
+                listener.onFail(-1,"local.m3u8 not exist。 请先点击解析 M3U8")
+                return
+            }
+            val finalOutPutFile = File(videoName)
             if (finalOutPutFile.exists()) {
                 finalOutPutFile.delete()
             }
             FileUtils.checkAndCreateFolder(finalOutPutFile.parentFile.absolutePath)
 
-            var a = parseIndex("", "", m3u8Dir + "local.m3u8")
             val fileOutputStream = FileOutputStream(finalOutPutFile, true)
             for (i in 0 until a.tsList.size) {
                 try {
@@ -224,7 +228,7 @@ object M3U8Tools {
             }
         } catch (e: java.lang.Exception) {
             e.printStackTrace()
-            listener.onFail(-1, e.toString())
+            listener.onFail(-2, e.toString())
             return
         }
         return listener.onComplete()
